feat: クエスト解放条件、FirstEventDragon 島破壊イベント調整、ロード画面演出、物理の固定ステップ安定化など
- クエスト解放条件をインターフェース + 合成 (Quest/Unlock) で指定できるようにし、クエストボードに反映
- FirstEventDragon: 島破壊イベントのパーティクル削減・会話タイミング/飛行カメラ修正、StartChat アクション追加、体力 2 倍・咆哮確率 1/2、咆哮 SE のタイミングと二重再生を修正
- ロード画面: ルートマップの演出を Tween 化、行き先の〇を描くアニメ、ロード BGM 追加
- SwordMan のジャンプ攻撃で敵の方を向かないように、攻撃時のカメラシェイク強化と MagicCaster の魔法にもシェイク追加
- GameCursor を Button 表示中だけ出し、押下/離しで伸縮するように
- RestorationGate が復興後のプレハブを子に生成するように (v2)、ストーリー / NPC 会話の文章を復元・更新
- Time を µs 計測に、RigidBody の Transform 書き戻しを物理ステップ直後にまとめて行うように
- コイン取得音をリアル寄りに作り直し、ハイエナの SE、FirstTouchDown の飛行船コライダーを Boundary レイヤーに
- tools.scene の catalog/edits 修正、/commit-push・/bugfix-recently-session スキル追加
- ログ・ImGui レイアウト・.meta を更新

+++ Engine/Core/Application/Time/Time.h
@@ -18,7 +18,7 @@ namespace NanamiEngine
         static void  SetFixedDeltaTime(float fixedDeltaTime);
 
     private:
-        static int lastTime_;
+        static long long lastTime_;
         static float deltaTime_;
         static float timeScale_;
         static float currentTime_;
